ALL FILES UNSORTED/requirements.txt
# --- requirements.txt ---
crewai
crewai_tools
langchain-community
ollama
psycopg2-binary
pgvector
bcrypt
cryptography
"requests[socks]"
selenium
numpy
scipy
pandas
scikit-learn
pyautogui
openai-whisper
openai
anthropic
twilio
pyotp
qrcode[pil]
GitPython
websocket-client
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart